regen

+++ sdk/resourcemanager/azure-resourcemanager-appservice/src/main/java/com/azure/resourcemanager/appservice/models/FunctionsAlwaysReadyConfig.java
@@ -13,16 +13,18 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @Fluent
 public final class FunctionsAlwaysReadyConfig {
     /*
-     * Either a function group or a function name is required. For additional information see https://aka.ms/flexconsumption/alwaysready.
+     * Either a function group or a function name is required. For additional information see
+     * https://aka.ms/flexconsumption/alwaysready.
      */
     @JsonProperty(value = "name")
     private String name;
 
     /*
-     * Sets the number of 'Always Ready' instances for a given function group or a specific function. For additional information see https://aka.ms/flexconsumption/alwaysready.
+     * Sets the number of 'Always Ready' instances for a given function group or a specific function. For additional
+     * information see https://aka.ms/flexconsumption/alwaysready.
      */
     @JsonProperty(value = "instanceCount")
-    private Float instanceCount;
+    private Integer instanceCount;
 
     /**
      * Creates an instance of FunctionsAlwaysReadyConfig class.
@@ -58,7 +60,7 @@ public final class FunctionsAlwaysReadyConfig {
      * 
      * @return the instanceCount value.
      */
-    public Float instanceCount() {
+    public Integer instanceCount() {
         return this.instanceCount;
     }
 
@@ -69,7 +71,7 @@ public final class FunctionsAlwaysReadyConfig {
      * @param instanceCount the instanceCount value to set.
      * @return the FunctionsAlwaysReadyConfig object itself.
      */
-    public FunctionsAlwaysReadyConfig withInstanceCount(Float instanceCount) {
+    public FunctionsAlwaysReadyConfig withInstanceCount(Integer instanceCount) {
         this.instanceCount = instanceCount;
         return this;
     }
